feat: google and apple login widget
