+++ lib/widgets/widgets.dart
@@ -3,5 +3,5 @@ library widgets;
 export 'appbar.dart';
 export 'bottomsheet.dart';
 export 'profile.dart';
-export 'googleLogin.dart';
+export 'logins.dart';
 export 'dialog.dart';
